Move classes to the same place
